API/requirements.txt
fastapi
uvicorn
pydantic
ultralytics
opencv-python-headless
numpy
google-cloud-storage